requirements.txt
alembic==1.8.0
APScheduler==3.9.1
astroid==2.11.7
black==22.1.0
blue==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colorlog==6.6.0
dill==0.3.5.1
flake8==4.0.1
greenlet==1.1.2
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
pathspec==0.9.0
pika==1.3.0
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.15.0
pyflakes==2.4.0
pylint==2.14.4
python-ldap==3.4.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
sqlalchemy-stubs==0.4
sqlalchemy2-stubs==0.0.2a24
tomli==2.0.1
tomlkit==0.11.1
types-requests==2.28.0
types-urllib3==1.26.16
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
wrapt==1.14.1
